+++ src/Core/Content/SalesChannelSnippet/SalesChannelSnippetDefinition.php
@@ -5,6 +5,7 @@ namespace SalesChannelSnippets\Core\Content\SalesChannelSnippet;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\CreatedAtField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\FkField;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\ApiAware;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\PrimaryKey;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\Required;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\IdField;
@@ -38,15 +39,15 @@ class SalesChannelSnippetDefinition extends EntityDefinition
     protected function defineFields(): FieldCollection
     {
         return new FieldCollection([
-            (new IdField('id', 'id'))->addFlags(new Required(), new PrimaryKey()),
-            (new FkField('sales_channel_id', 'salesChannelId', SalesChannelDefinition::class))->addFlags(new Required()),
-            (new FkField('language_id', 'languageId', LanguageDefinition::class))->addFlags(new Required()),
-            (new StringField('translation_key', 'translationKey'))->addFlags(new Required()),
-            (new LongTextField('value', 'value'))->addFlags(new Required()),
-            new ManyToOneAssociationField('salesChannel', 'sales_channel_id', SalesChannelDefinition::class, 'id', false),
-            new ManyToOneAssociationField('language', 'language_id', LanguageDefinition::class, 'id', false),
-            new CreatedAtField(),
-            new UpdatedAtField(),
+            (new IdField('id', 'id'))->addFlags(new Required(), new PrimaryKey(), new ApiAware()),
+            (new FkField('sales_channel_id', 'salesChannelId', SalesChannelDefinition::class))->addFlags(new Required(), new ApiAware()),
+            (new FkField('language_id', 'languageId', LanguageDefinition::class))->addFlags(new Required(), new ApiAware()),
+            (new StringField('translation_key', 'translationKey'))->addFlags(new Required(), new ApiAware()),
+            (new LongTextField('value', 'value'))->addFlags(new Required(), new ApiAware()),
+            (new ManyToOneAssociationField('salesChannel', 'sales_channel_id', SalesChannelDefinition::class, 'id', false))->addFlags(new ApiAware()),
+            (new ManyToOneAssociationField('language', 'language_id', LanguageDefinition::class, 'id', false))->addFlags(new ApiAware()),
+            (new CreatedAtField())->addFlags(new ApiAware()),
+            (new UpdatedAtField())->addFlags(new ApiAware()),
         ]);
     }
 }